Moves EnvData source and header files to /src and /include
