+++ src/disturb/WildFire.h
@@ -7,7 +7,7 @@
 #include <algorithm>
 
 #include "../../include/CohortData.h"
-#include "../data/EnvData.h"
+#include "../../include/EnvData.h"
 #include "../data/FirData.h"
 #include "../../include/BgcData.h"
 #include "../data/RestartData.h"
