+++ src/snowsoil/Snow_Env.h
@@ -4,7 +4,7 @@
 #include "../inc/errorcode.h"
 
 #include "../../include/CohortData.h"
-#include "../data/EnvData.h"
+#include "../../include/EnvData.h"
 #include "../data/RestartData.h"
 
 #include "../ecodomain/Ground.h"
