+++ src/snowsoil/Soil_Bgc.h
@@ -4,7 +4,7 @@
 #include "../inc/errorcode.h"
 
 #include "../../include/CohortData.h"
-#include "../data/EnvData.h"
+#include "../../include/EnvData.h"
 #include "../data/FirData.h"
 #include "../../include/BgcData.h"
 #include "../data/RestartData.h"
