+++ src/snowsoil/Soil_Env.h
@@ -6,7 +6,7 @@
 #include "TemperatureUpdator.h"
 
 #include "../../include/CohortData.h"
-#include "../data/EnvData.h"
+#include "../../include/EnvData.h"
 #include "../data/FirData.h"
 #include "../data/RestartData.h"
