+++ src/vegetation/Vegetation_Env.h
@@ -9,7 +9,7 @@
 #include "../inc/parameters.h"
 
 #include "../../include/CohortData.h"
-#include "../data/EnvData.h"
+#include "../../include/EnvData.h"
 #include "../data/FirData.h"
 #include "../data/RestartData.h"
